Refactor the gui code
