+++ src/main/java/Duke.java
@@ -3,54 +3,35 @@ import java.util.Scanner;
 import command.Command;
 import dukeexeption.DukeException;
 import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.Region;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 import parser.Request;
 import storage.LocalStorage;
 import storage.TaskList;
 import task.Task;
-import ui.DialogBox;
-import ui.Ui;
+import ui.Gui;
+import ui.TextUi;
 
 /**
  * The main duke program.
  */
 public class Duke extends Application {
     private TaskList tasks;
-    private Ui ui;
+    private TextUi textUi;
     private LocalStorage localTaskList;
 
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-
-    private Image user = new Image(this.getClass().getResourceAsStream("/images/DaUser.png"));
-    private Image duke = new Image(this.getClass().getResourceAsStream("/images/DaDuke.png"));
-
     public Duke() {
-        this.ui = new Ui();
+        this.textUi = new TextUi();
         this.tasks = new TaskList();
     }
 
     public Duke(String filepath) {
-        this.ui = new Ui();
+        this.textUi = new TextUi();
         ArrayList<Task> taskList = new ArrayList<>();
         try {
             this.localTaskList = new LocalStorage(filepath);
             taskList = this.localTaskList.createTaskList();
         } catch (DukeException error) {
-            ui.printFormattedError(error);
+            textUi.printFormattedError(error);
         }
         this.tasks = new TaskList(taskList);
     }
@@ -63,7 +44,7 @@ public class Duke extends Application {
      * Begins the execution of the Duke program.
      */
     public void run() {
-        this.ui.printStartUpMessage();
+        this.textUi.printStartUpMessage();
 
         Scanner scanner = new Scanner(System.in);
 
@@ -78,120 +59,31 @@ public class Duke extends Application {
             try {
                 Command command = new Request(request).parse();
                 String reply = command.run(this.tasks);
-                this.ui.printFormattedResponse(reply);
+                this.textUi.printFormattedResponse(reply);
             } catch (DukeException error) {
-                this.ui.printFormattedError(error);
+                this.textUi.printFormattedError(error);
             }
         }
 
-        this.ui.printExitingMessage();
+        this.textUi.printExitingMessage();
     }
 
     @Override
     public void start(Stage stage) {
-        //Step 1. Setting up required components
-
-        //The container for the content of the chat to scroll.
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
-
-        userInput = new TextField();
-        sendButton = new Button("Send");
-
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-        stage.show();
-
-        //Step 2. Formatting the window to look as expected
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-
-        mainLayout.setPrefSize(400.0, 600.0);
-
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        // You will need to import `javafx.scene.layout.Region` for this.
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(325.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput , 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-
-        //Step 3. Add functionality to handle user input.
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
-
-        //Scroll down to the end every time dialogContainer's height changes.
-        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-
-        // more code to be added here later
-    }
-
-    /**
-     * Iteration 1:
-     * Creates a label with the specified text and adds it to the dialog container.
-     * @param text String containing text to add
-     * @return a label with the specified text that has word wrap enabled.
-     */
-    private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
-
-        return textToAdd;
-    }
-
-    /**
-     * Iteration 2:
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    private void handleUserInput() {
-        Label userText = new Label(userInput.getText());
-        Label dukeText = new Label(getResponse(userInput.getText()));
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(userText, new ImageView(user)),
-                DialogBox.getDukeDialog(dukeText, new ImageView(duke))
-        );
-        userInput.clear();
+        new Gui((String input) -> this.parseAndRespond(input)).start(stage);
     }
 
     /**
      * You should have your own function to generate a response to user input.
      * Replace this stub with your completed method.
      */
-    private String getResponse(String input) {
+    private String parseAndRespond(String input) {
         try {
             Command command = new Request(input).parse();
             String reply = command.run(this.tasks);
             return "Duke shouts:\n " + reply;
         } catch (DukeException error) {
-            return "Duke shouts:\n " + error.toString();
+            return "Duke shouts:\n " + error;
         }
     }
 }
